azure-function/requirements.txt
azure-functions
requests==2.31.0
openai==1.14.0
python-dotenv==1.0.0
beautifulsoup4>=4.9.3,<5.0.0
msal~=1.20
